refactor: extract run_event_loop into handler modules (#437)
Split the 1,673-line tui_context.rs into 4 focused files:

- tui_context.rs (400 lines) — struct + new/draw/emit/cleanup + thin
  run_event_loop dispatcher
- tui_handlers_commands.rs (383 lines) — slash command routing, dropdown
  openers (/model, /provider, /sessions), inference start preparation
- tui_handlers_idle.rs (560 lines) — idle keyboard/paste/menu navigation,
  wizard submit, history, reactive menus
- tui_handlers_inference.rs (480 lines) — inference turn inner loop,
  approval/loop-cap hotkeys, UI event rendering, post-turn cleanup +
  auto-compact

Key design decisions:
- run_event_loop() is now a ~40-line dispatcher calling handler methods
- Inference inner loop uses free functions (handle_inference_key_inline,
  handle_inference_ui_inline) with field-level borrows to satisfy the
  borrow checker while the pinned turn future holds &mut session
- save_history promoted to pub(crate) for cross-module access
- All files under 600 lines
- Zero warnings, all tests pass
