+++ koda-cli/src/main.rs
@@ -17,6 +17,9 @@ mod tool_history;
 mod tui_app;
 mod tui_commands;
 mod tui_context;
+mod tui_handlers_commands;
+mod tui_handlers_idle;
+mod tui_handlers_inference;
 mod tui_output;
 mod tui_render;
 mod tui_types;
